Add login decorators to the relevant views

+++ courses/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
-from django.db.models import Q
-from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
+
 from .models import Course
 
 
@@ -32,8 +32,12 @@ def course_detail(request, course_id):
     return render(request, 'courses/course_detail.html', context)
 
 
+@login_required
 def add_course(request):
     """ Add a course to the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES)
@@ -42,10 +46,11 @@ def add_course(request):
             messages.success(request, 'Successfully added Course!')
             return redirect(reverse('add_course'))
         else:
-            messages.error(request, 'Failed to add course. Please ensure the form is valid.')
+            messages.error(
+                request, 'Failed to add course. Please ensure the form is valid.')
     else:
         form = CourseForm()
-        
+
     template = 'courses/add_course.html'
     context = {
         'form': form,
@@ -54,8 +59,13 @@ def add_course(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_course(request, course_id):
     """ Edit a course in the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES, instance=course)
@@ -64,7 +74,8 @@ def edit_course(request, course_id):
             messages.success(request, 'Successfully updated course!')
             return redirect(reverse('course_detail', args=[course.id]))
         else:
-            messages.error(request, 'Failed to update course. Please ensure the form is valid.')
+            messages.error(
+                request, 'Failed to update course. Please ensure the form is valid.')
     else:
         form = CourseForm(instance=course)
         messages.info(request, f'You are editing {course.name}')
@@ -78,8 +89,13 @@ def edit_course(request, course_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_course(request, course_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     course.delete()
     messages.success(request, 'Course deleted!')
