+++ exam_courses/views.py
@@ -1,7 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
-from django.db.models import Q
-from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 from .models import ExamCourse
 
 from .forms import ExamCourseForm
@@ -31,8 +30,13 @@ def exam_course_detail(request, exam_course_id):
     return render(request, 'exam_courses/exam_course_detail.html', context)
 
 
+@login_required
 def add_exam_course(request):
     """ Add a course to the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ExamCourseForm(request.POST, request.FILES)
         if form.is_valid():
@@ -51,8 +55,13 @@ def add_exam_course(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_exam_course(request, exam_course_id):
     """ Edit a course in the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     exam_course = get_object_or_404(ExamCourse, pk=exam_course_id)
     if request.method == 'POST':
         form = ExamCourseForm(request.POST, request.FILES, instance=exam_course)
@@ -75,8 +84,13 @@ def edit_exam_course(request, exam_course_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_exam_course(request, exam_course_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     exam_course = get_object_or_404(ExamCourse, pk=exam_course_id)
     exam_course.delete()
     messages.success(request, 'Course deleted!')
